==> notebooks/01_EDA_and_feature_engineering.ipynb <==
import pickle

import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder, StandardScaler

!pip install --upgrade keras

import keras

print(keras.__version__)

## Load the dataset
data = pd.read_csv(
    "/Users/sunnythesage/PythonProjects/Data-Science-BootCamp/03-Deep-Learning-BootCamp/7 - End to End Deep Learning Project Using ANN/advanced-customer-churn-analysis-using-ann/data/raw/churn-modelling-dataset.csv")

data.head()

### Data Preprocessing

# rop irrelevant columns

data = data.drop(['RowNumber', 'CustomerId', 'Surname'], axis = 1)
data

## Encode categorical variables

label_encoder_gender = LabelEncoder()

data['Gender'] = label_encoder_gender.fit_transform(data['Gender'])
data

## Onehot encode Geography column
from sklearn.preprocessing import OneHotEncoder

onehot_encoder_geo = OneHotEncoder()
geo_encoder = onehot_encoder_geo.fit_transform(data[['Geography']]).toarray()
geo_encoder

onehot_encoder_geo.get_feature_names_out(['Geography'])

geo_encoded_df = pd.DataFrame(geo_encoder, columns = onehot_encoder_geo.get_feature_names_out(['Geography']))
geo_encoded_df

## Combine one hot encoder columns with the original data
data = pd.concat([data.drop('Geography', axis = 1), geo_encoded_df], axis = 1)
data.head()

# Save encoders and scaler for later use

import os

# Define the project's root directory
project_root = "/Users/sunnythesage/PythonProjects/Data-Science-BootCamp/03-Deep-Learning-BootCamp/7 - End to End Deep Learning Project Using ANN/advanced-customer-churn-analysis-using-ann"

# Change the current working directory to the project's root
os.chdir(project_root)

# --- Artifacts ---

# Define the relative path to the artifacts directory
artifacts_dir = os.path.join(os.getcwd(), 'artifacts')

# Create the directory if it doesn't exist
os.makedirs(artifacts_dir, exist_ok = True)

print(artifacts_dir)

# Save encoders and scaler to the artifacts folder
with open(os.path.join(artifacts_dir, 'label_encoder_gender.pkl'), 'wb') as file:
    pickle.dump(label_encoder_gender, file)

with open(os.path.join(artifacts_dir, 'onehot_encoder_geo.pkl'), 'wb') as file:
    pickle.dump(onehot_encoder_geo, file)

data.head()

## Divide the dataset into independent and dependent features
X = data.drop('Exited', axis = 1)
y = data['Exited']

## Split the data in train and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

## Scale these features
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

X_train

with open(os.path.join(artifacts_dir, 'scaler.pkl'), 'wb') as file:
    pickle.dump(scaler, file)

data

### ANN Implementation

from keras import layers

from keras import Sequential

(X_train.shape[1],)

# Build Our ANN Model
model = Sequential([
    layers.Input(shape = (X_train.shape[1],)),  # Explicitly define the input shape
    layers.Dense(64, activation = 'relu'),  # HL1
    layers.Dense(32, activation = 'relu'),  # HL2
    layers.Dense(1, activation = 'sigmoid')  # Output layer
])

model.summary()

opt = keras.optimizers.Adam(learning_rate = 0.01)
loss = keras.losses.BinaryCrossentropy()
loss

## compile the model
model.compile(optimizer = opt, loss = "binary_crossentropy", metrics = ['accuracy'])

## Set up the Tensorboard

import datetime

# --- Logs ---

# Define the root log directory (similar to artifacts)
root_log_dir = os.path.join(os.getcwd(), 'logs')

# Create the root log directory if it doesn't exist
os.makedirs(root_log_dir, exist_ok = True)

# Generate a unique log directory for each run
log_dir = os.path.join(root_log_dir, "fit", datetime.datetime.now().strftime("%Y%m%d-%H%M%S"))

# Create the specific log directory for this run
os.makedirs(log_dir, exist_ok = True)

print(root_log_dir)
print(log_dir)

# Create the TensorBoard callback
tensorflow_callback = keras.callbacks.TensorBoard(log_dir = log_dir, histogram_freq = 1)

## Set up Early Stopping
early_stopping_callback = keras.callbacks.EarlyStopping(monitor = 'val_loss', patience = 10,
                                                        restore_best_weights = True)


### Train the model
history = model.fit(
    X_train, y_train, validation_data = (X_test, y_test), epochs = 100,
    callbacks = [tensorflow_callback, early_stopping_callback]
)

# model.save('model.keras')

keras.saving.save_model(model, os.path.join(artifacts_dir, 'model.keras'))

## Load Tensorboard Extension
%load_ext tensorboard

%tensorboard --logdir logs/fit
