==> notebooks/02_prediction.ipynb <==
import os
import pickle

import pandas as pd

!pip install --upgrade keras

import keras

print(keras.__version__)

# Define the project's root directory (replace with your actual path)
project_root = "/Users/sunnythesage/PythonProjects/Data-Science-BootCamp/03-Deep-Learning-BootCamp/7 - End to End Deep Learning Project Using ANN/advanced-customer-churn-analysis-using-ann"

# Change the current working directory to the project's root
os.chdir(project_root)

# Define the artifacts directory path
artifacts_dir = os.path.join(os.getcwd(), 'artifacts')

# Load the trained model
model_path = os.path.join(artifacts_dir, 'model.keras')  # use .keras
model = keras.saving.load_model(model_path)

# Load the encoder and scaler
with open(os.path.join(artifacts_dir, 'onehot_encoder_geo.pkl'), 'rb') as file:
    onehot_encoder_geo = pickle.load(file)

with open(os.path.join(artifacts_dir, 'label_encoder_gender.pkl'), 'rb') as file:
    label_encoder_gender = pickle.load(file)

with open(os.path.join(artifacts_dir, 'scaler.pkl'), 'rb') as file:
    scaler = pickle.load(file)

# Example input data
input_data = {
    'CreditScore': 600,
    'Geography': 'France',
    'Gender': 'Male',
    'Age': 40,
    'Tenure': 3,
    'Balance': 60000,
    'NumOfProducts': 2,
    'HasCrCard': 1,
    'IsActiveMember': 1,
    'EstimatedSalary': 50000
}

# One-hot encode 'Geography'
# Create a DataFrame with the 'Geography' value
geo_df = pd.DataFrame([input_data['Geography']], columns = ['Geography'])

# One-hot encode 'Geography'
geo_encoded = onehot_encoder_geo.transform(geo_df).toarray()

# Create a DataFrame from the encoded data
geo_encoded_df = pd.DataFrame(geo_encoded, columns = onehot_encoder_geo.get_feature_names_out(['Geography']))

geo_encoded_df


input_df = pd.DataFrame([input_data])
input_df

## Encode categorical variables
input_df['Gender'] = label_encoder_gender.transform(input_df['Gender'])
input_df

## concatenation one hot encoded
input_df = pd.concat([input_df.drop("Geography", axis = 1), geo_encoded_df], axis = 1)
input_df

## Scaling the input data
input_scaled = scaler.transform(input_df)
input_scaled

## PRedict churn
prediction = model.predict(input_scaled)
prediction

prediction_proba = prediction[0][0]

prediction_proba

if prediction_proba > 0.5:
    print('The customer is likely to churn.')
else:
    print('The customer is not likely to churn.')
