==> notebooks/03_salary_regression_implementation.ipynb <==
import datetime
import os

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder, OneHotEncoder, StandardScaler

# Define the project's root directory (replace with your actual path)
project_root = "/Users/sunnythesage/PythonProjects/Data-Science-BootCamp/03-Deep-Learning-BootCamp/7 - End to End Deep Learning Project Using ANN/advanced-customer-churn-analysis-using-ann"

# Change the current working directory to the project's root
os.chdir(project_root)

# Define the artifacts directory path
artifacts_dir = os.path.join(os.getcwd(), 'artifacts')

# Create the artifacts directory if it doesn't exist
os.makedirs(artifacts_dir, exist_ok = True)

## Load the dataset
data = pd.read_csv('data/raw/churn-modelling-dataset.csv')

data.head()

# Preprocess the data
data = data.drop(['RowNumber', 'CustomerId', 'Surname'], axis = 1)

# Encode categorical variables
label_encoder_gender = LabelEncoder()
data['Gender'] = label_encoder_gender.fit_transform(data['Gender'])

data

# One-hot encode 'Geography'
onehot_encoder_geo = OneHotEncoder(handle_unknown = 'ignore')

geo_encoded = onehot_encoder_geo.fit_transform(data[['Geography']]).toarray()
geo_encoded_df = pd.DataFrame(geo_encoded, columns = onehot_encoder_geo.get_feature_names_out(['Geography']))

geo_encoded_df

# Combine one-hot encoded column with original data
data = pd.concat([data.drop('Geography', axis = 1), geo_encoded_df], axis = 1)
data.head()

# Split the data into features and target
X = data.drop('EstimatedSalary', axis = 1)
y = data['EstimatedSalary']

## Split the data in training and testing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

## Scale these features
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

#### ANN Regression Problem statement

import keras
from keras import layers

# Build Our ANN Model
model = keras.Sequential([
    layers.Input(shape = (X_train.shape[1],)),  # Explicitly define the input shape
    layers.Dense(64, activation = 'relu'),  # HL1
    layers.Dense(32, activation = 'relu'),  # HL2
    layers.Dense(1, activation = 'sigmoid')  # Output layer
])

# Compile the model
opt = keras.optimizers.Adam(learning_rate = 0.01)
model.compile(optimizer = opt, loss = 'binary_crossentropy', metrics = ['accuracy'])

model.summary()

# Define the root log directory
root_log_dir = os.path.join(os.getcwd(), 'logs')
os.makedirs(root_log_dir, exist_ok = True)

# Generate a unique log directory for each run
log_dir = os.path.join(root_log_dir, 'fit', datetime.datetime.now().strftime('%Y%m%d-%H%M%S'))
os.makedirs(log_dir, exist_ok = True)

# Set up TensorBoard
log_dir = "regression_logs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")

# Create the TensorBoard callback
tensorboard_callback = keras.callbacks.TensorBoard(log_dir = log_dir, histogram_freq = 1)

# Set up Early Stopping
early_stopping_callback = keras.callbacks.EarlyStopping(monitor = 'val_loss', patience = 10,
                                                        restore_best_weights = True)

# Train the model
history = model.fit(
    X_train, y_train,
    validation_data = (X_test, y_test),
    epochs = 100,
    callbacks = [tensorboard_callback, early_stopping_callback]
)

## Evaluate model on the test data
test_loss, test_mae = model.evaluate(X_test, y_test)

print(f'Test MAE : {test_mae}')

# Save the model in the artifacts directory with .keras extension
keras.saving.save_model(model, os.path.join(artifacts_dir, 'regression_model.keras'))

# Load Tensorboard Extension
%load_ext tensorboard

%tensorboard --logdir logs/fit
